requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

adjustText==0.8
albumentations==1.3.1
einops==0.7.0
etils==1.9.1
loguru==0.7.2
matplotlib==3.8.2
mlcroissant==1.0.5
mmengine==0.10.2
numpy==1.26.2
open-clip-torch==2.24.0
opencv-python-headless==4.8.1.78
Pillow==10.0.1
pysodmetrics==1.4.2
PyYAML==6.0
timm==0.9.12
# Editable install with no version control (torch==2.1.2)
# -e e:\programming\miniconda\envs\pt212\lib\site-packages\torch-2.1.2-py3.11.egg-info
# Editable install with no version control (torchvision==0.16.2)
# -e e:\programming\miniconda\envs\pt212\lib\site-packages\torchvision-0.16.2-py3.11.egg-info
tqdm==4.66.1
